Update requirements for Debian and Fedora
The requirements were lacking packages needed for successful completion
of the script, including gnome-shell. I have also added the new
requirements I've brought in: bc and xml_grep.

bc should be already installed on every system as part of the base
install, so should be entirely non-controversial for anyone looking at
the requirements list. It is required as shell math is integer and X.Y
style versioning can be treated as a floating point trivially.

xml_grep could be treated as more of a "soft" requirement, with a slight
logic change to utils.sh. It ensures that the correct version number of
the overarching GNOME install is detected, rather than the
potentially-different GNOME Shell version.